% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVI.R
\name{calcVI}
\alias{calcVI}
\title{vegetion index calculation handler}
\usage{
calcVI(index, spc, sensor = "SVC", ...)
}
\arguments{
\item{index:}{name of index, used to match function}

\item{spc:}{spc obj}

\item{sensor:}{in c('SVC', 'SEN', 'LS8') to match bands}
}
\value{
vi vector
}
\description{
vegetion index calculation handler
}
