% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun4mutate.R
\name{index2factor}
\alias{index2factor}
\title{change index to factor for display}
\usage{
index2factor(x)
}
\arguments{
\item{x}{index vector}
}
\value{
x factored with labels for plotmath
}
\description{
change index to factor for display
}
