% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun4mutate.R
\name{index2latex}
\alias{index2latex}
\title{change index  to for latex}
\usage{
index2latex(x)
}
\arguments{
\item{x}{index vector in factor mode}
}
\value{
x factored with labels for latex
}
\description{
change index  to for latex
}
