% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun4mutate.R
\name{lowCor2NA}
\alias{lowCor2NA}
\title{change unsignificant cor value to NA according to pvalue}
\usage{
lowCor2NA(pvalue, estimate, thresh = 0.01)
}
\arguments{
\item{pvalue}{pvalue}

\item{estimate}{cor value}

\item{thresh}{pvalue thresh}
}
\value{
update estimate
}
\description{
change unsignificant cor value to NA according to pvalue
}
