% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelInfo_kerasL1.R
\name{modelInfo_kerasL1}
\alias{modelInfo_kerasL1}
\title{model info for caret to do keras fitting}
\usage{
modelInfo_kerasL1()
}
\value{
list
}
\description{
model info for caret to do keras fitting
}
