% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4raster.R
\name{raster_levelplot}
\alias{raster_levelplot}
\title{FIXME: demo}
\usage{
raster_levelplot(fpath_in, title, noData = -9999)
}
\arguments{
\item{fpath_in}{fpath}

\item{title}{plot title}

\item{noData}{noDate value default -9999}
}
\description{
FIXME: demo
}
