% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun4mutate.R
\name{RMSE2rRMSE}
\alias{RMSE2rRMSE}
\title{change RMSE to normalized RMSE}
\usage{
RMSE2rRMSE(tag, RMSEvec, obsMean)
}
\arguments{
\item{tag:}{tag vector to indicate each group(not factor)}

\item{RMSEvec:}{RMSE vector}

\item{obsMean:}{observed mean vector(must have name to match with tag)}
}
\value{
nRMSE in %
}
\description{
change RMSE to normalized RMSE
}
