% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_cor}
\alias{spc_cor}
\title{Calc r between reflectance and one of SI}
\usage{
spc_cor(spc, biochemphy, method = "pearson", use = "complete.obs")
}
\arguments{
\item{spc:}{spc obj}

\item{biochemphy:}{name of SI}
}
\value{
tibble with estiamte and pvalue
}
\description{
Calc r between reflectance and one of SI
}
