% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4plot.R
\name{spc_plot}
\alias{spc_plot}
\title{plot speclib}
\usage{
spc_plot(x, FUN = NULL)
}
\arguments{
\item{x}{spc obj}

\item{FUN}{ids of spectral(vector) or NULL}
}
\value{
ggplot
}
\description{
plot speclib
}
